test(article): fold cases into main_content suite
Move every article() case into tests/serialize/test_tree_main_content.py as
shared-fixture parametrized cases and drop the standalone file; article builds
on the same readability heuristic as main_content.
